Display ground texture and intensity

+++ src/island/layout.js
@@ -13,7 +13,7 @@ export function loadLayout(ile) {
                 x: (16 - x) - 8,
                 y: y - 8,
                 triangles: new Uint32Array(ile.getEntry(id * 6 - 1)),
-                textureInfo: new Uint16Array(ile.getEntry(id * 6)),
+                textureInfo: new Uint8Array(ile.getEntry(id * 6)),
                 heightmap: new Uint16Array(ile.getEntry(id * 6 + 1)),
                 intensity: new Uint8Array(ile.getEntry(id * 6 + 2))
             });
